Extract context selector button
